remove hard-coded db connection credentials
The mongolab db user creditials were hard-coded in the developmeent.js
config file, which was accessible to the public on github.  Even though
this is code only follows along with the book, I thought it would be a
good idea to refactor the credentials to environment variables and
source those to the operating system.

The db user creditials have been moved to a separate file which has been
added to .gitignore