modify the user's routes for authentication views
This wires up the user's routes for the signin and signup pages.  Also
fixes a few minor oversights in previous commits (from previous
sections) to ensure the server runs error free.

To the point the code is tested (from a functional perspective, not unit
tested) and all functionality works as advertised.