add facebook oauth authentication strategy
This adds the ability to be able to login using OAuth through facebook.
For some reason though, the profile sent back from facebook does not
include the profile.emails array (at least in my case) which is used to
pull the user email address for the user model.

However, this does appear to be working when using a similar approach
with the scotch.io tutorial.  More research is required.