fix facebook oauth authentication strategy
Facebook was not sending back the user's email, which was required to
build up the local user model, as part of the returned profile.  This
was because we were not asking.  Needed to include a 'scope' property to
the request in the route.  This was not included in the code from the
book, some research was required.