Finish code for chapter 7
The code in the book calls for a separate file to define the 'users'
angular module then list in index.ejs, which makes sense.  However, when
that is done, the system shows a 500 internal server error with some
information in the stacktrace about a mongoose cast error, which makes
no sense.  Why does mongoose, or mongodb, have anything to do with
sourcing a simple javascript file in an html page.

This issue is resolved however when the angular module definition is
moved to the service (factory) file in a subdirectory.  I don't get it...