fix user module loading issue
This fixes the error 500 issue with loading the users.client.module.js
file from the public/users directory.  For some unknown reason, it would
not load the module file when the directory is named 'users'.  Renamed
directory to 'users-module' and it worked fine.

There seemed to be some issue/conflict with mongoose.  Issue is resolved
and not worth the effort at this point to determine root cause...moving on.